fix: use Math.toIntExact for pageable offset to prevent int overflow
Agent-Logs-Url: https://github.com/windrunner0707/transaction-system/sessions/1c8001bc-d8d5-4cf8-a851-4ccdaf38dc9a

+++ src/main/java/com/qiqibai/transactionsystem/infrastructure/InMemoryTransactionRepository.java
@@ -65,7 +65,7 @@ public class InMemoryTransactionRepository implements TransactionRepository {
                 .map(Transaction::copy)
                 .toList();
         int total = all.size();
-        int start = (int) pageable.getOffset();
+        int start = Math.toIntExact(pageable.getOffset());
         if (start >= total) {
             return new PageImpl<>(List.of(), pageable, total);
         }
